Fixed some past events slipping through
Error was the date was initially in the format of DD/MM/YYYY so when it was comparing to todays date it was the condition was correct up till a day was above the current day. So changed the date to YYYY/MM/DD so after checking the condition for the Year it would check the Month then Day.